Moving semi honest code inside its own folder
